VEOSS-1360 | Move remaining previous reg files into correct directory
